feat: Remove setting replaceMockedFetches
BREAKING CHANGE: This shouldn't break anything, because this setting was
used only to validate adding new mock `fetch` handlers, so that the same
handler wouldn't be added twice. But the understanding of "the same" may
vary. Does it mean the same URL pattern? Or a URL pattern matching the
same URL? Better to remove it.